requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
asyncio==3.4.3
attrs==23.2.0
azure-ai-vision-imageanalysis==1.0.0b2
azure-core==1.30.1
binaryornot==0.4.4
bitarray==2.9.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
exceptiongroup==1.2.0
fastapi==0.110.0
flet==0.21.2
flet-core==0.21.2
flet-runtime==0.21.2
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
isodate==0.6.1
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
pdf2image==1.17.0
pdfreader==0.1.12
pillow==10.2.0
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
PyPDF2==3.0.1
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
qrcode==7.4.2
repath==0.9.0
requests==2.31.0
rich==13.7.1
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.36.3
text-unidecode==1.3
trio==0.25.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchdog==4.0.0
watchfiles==0.21.0
webdriver-manager==4.0.1
websockets==12.0
wsproto==1.2.0
zipp==3.18.1
